about page

+++ client/src/pages/Auth.jsx
@@ -3,6 +3,7 @@ import { useState } from "react";
 import "../style/auth.css";
 import { useNavigate } from "react-router-dom";
 import { useCookies } from "react-cookie";
+import Swal from 'sweetalert2'
 
 export const Auth = () => {
   const [isLogin, setIsLogin] = useState(true);
@@ -35,13 +36,25 @@ const Login = ({ toggleForm }) => {
         username,
         password,
       });
-      setCookies("access_token", response.data.token);
-      window.localStorage.setItem("userID" , response.data.userID)
-      window.localStorage.setItem("username" , response.data.username)
-      alert(`Welcome Back ${response.data.username}`)
-      navigate('/')
+      if (response.data.message) {
+        Swal.fire({
+          title : "Error!",
+          text : response.data.message,
+          icon : 'error'
+        })
+      } else {
+        setCookies("access_token", response.data.token);
+        window.localStorage.setItem("userID" , response.data.userID)
+        window.localStorage.setItem("username" , response.data.username)
+        Swal.fire({
+          title : "Success",
+          text : `Welcome Back ${response.data.username.toUpperCase()}`,
+          icon : 'success'
+        })
+        navigate('/')
+      }
     } catch (err) {
-      console.log(err);
+      console.log(err.response.data.error);
     }
   };
 
@@ -67,14 +80,26 @@ const Register = ({ toggleForm }) => {
   const onSubmit = async (e) => {
     e.preventDefault();
     try {
-      await axios.post("http://localhost:3001/auth/register", {
+      const response = await axios.post("http://localhost:3001/auth/register", {
         username,
         password,
       });
-      alert("Registration completed! Now login");
-      toggleForm();
+      if (response.data.message !== "User already exists!") {
+        Swal.fire({
+          title : "Success",
+          text : "Registration completed! Now login",
+          icon : 'Success'
+        })
+        toggleForm();
+      } else {
+        Swal.fire({
+          title : "Error!",
+          text : response.data.message,
+          icon : 'error'
+        })
+      }
     } catch (err) {
-      console.log(err);
+      console.error(err);
     }
   };
   return (
